wip: start header/trailer cleanup, cf. #84
