+++ crates/hring/src/h2/end_headers.rs
@@ -1,0 +1,298 @@
+use std::{borrow::Cow, rc::Rc};
+
+use eyre::ErrReport;
+use hring_buffet::{Piece, PieceStr};
+use http::{
+    header,
+    uri::{Authority, PathAndQuery, Scheme},
+    HeaderName, Version,
+};
+use tokio::sync::mpsc;
+use tracing::{debug, warn};
+
+use crate::{ExpectResponseHeaders, Headers, Method, Request, Responder, ServerDriver};
+
+use super::{
+    body::{H2Body, H2BodyItem, H2BodySender, PieceOrTrailers},
+    encode::{EncoderState, H2ConnEvent, H2Encoder},
+    parse::{KnownErrorCode, StreamId},
+    send_goaway, HeadersData, StreamRxStage, StreamState,
+};
+
+pub(crate) fn end_headers(
+    ev_tx: &mpsc::Sender<H2ConnEvent>,
+    stream_id: StreamId,
+    stream_state: &mut StreamState,
+    data: &HeadersData,
+    driver: &Rc<impl ServerDriver + 'static>,
+    hpack_dec: &mut hring_hpack::Decoder,
+    headers_or_trailers: HeadersOrTrailers,
+) -> EndHeadersResult {
+    match end_headers_inner(
+        ev_tx,
+        stream_id,
+        stream_state,
+        data,
+        driver,
+        hpack_dec,
+        headers_or_trailers,
+    ) {
+        Ok(res) => res,
+        Err(err) => EndHeadersResult {
+            stream_id,
+            payload: EndHeadersResultPayload::Err(err),
+        },
+    }
+}
+
+fn end_headers_inner(
+    ev_tx: &mpsc::Sender<H2ConnEvent>,
+    stream_id: StreamId,
+    stream_state: &mut StreamState,
+    data: &HeadersData,
+    driver: &Rc<impl ServerDriver + 'static>,
+    hpack_dec: &mut hring_hpack::Decoder,
+    headers_or_trailers: HeadersOrTrailers,
+) -> eyre::Result<EndHeadersResult> {
+    let mut method: Option<Method> = None;
+    let mut scheme: Option<Scheme> = None;
+    let mut path: Option<PieceStr> = None;
+    let mut authority: Option<Authority> = None;
+
+    let mut headers = Headers::default();
+
+    // TODO: find a way to propagate errors from here - probably will have to change
+    // the function signature in hring-hpack.
+    let on_header_pair = |key: Cow<[u8]>, value: Cow<[u8]>| {
+        debug!(
+            "{headers_or_trailers:?} | {}: {}",
+            std::str::from_utf8(&key).unwrap_or("<non-utf8-key>"), // TODO: does this hurt performance when debug logging is disabled?
+            std::str::from_utf8(&value).unwrap_or("<non-utf8-value>"),
+        );
+
+        if &key[..1] == b":" {
+            if matches!(headers_or_trailers, HeadersOrTrailers::Trailers) {
+                // TODO: proper error handling
+                panic!("trailers cannot contain pseudo-headers");
+            }
+
+            // it's a pseudo-header!
+            // TODO: reject headers that occur after pseudo-headers
+            match &key[1..] {
+                b"method" => {
+                    // TODO: error handling
+                    let value: PieceStr = Piece::from(value.to_vec()).to_str().unwrap();
+                    if method.replace(Method::try_from(value).unwrap()).is_some() {
+                        unreachable!(); // No duplicate allowed.
+                    }
+                }
+                b"scheme" => {
+                    // TODO: error handling
+                    let value: PieceStr = Piece::from(value.to_vec()).to_str().unwrap();
+                    if scheme.replace(value.parse().unwrap()).is_some() {
+                        unreachable!(); // No duplicate allowed.
+                    }
+                }
+                b"path" => {
+                    // TODO: error handling
+                    let value: PieceStr = Piece::from(value.to_vec()).to_str().unwrap();
+                    if value.len() == 0 || path.replace(value).is_some() {
+                        unreachable!(); // No empty path nor duplicate allowed.
+                    }
+                }
+                b"authority" => {
+                    // TODO: error handling
+                    let value: PieceStr = Piece::from(value.to_vec()).to_str().unwrap();
+                    if authority.replace(value.parse().unwrap()).is_some() {
+                        unreachable!(); // No duplicate allowed. (h2spec doesn't seem to test for
+                                        // this case but rejecting for duplicates seems
+                                        // reasonable.)
+                    }
+                }
+                _ => {
+                    debug!("ignoring pseudo-header");
+                }
+            }
+        } else {
+            // TODO: what do we do in case of malformed header names?
+            // ignore it? return a 400?
+            let name = HeaderName::from_bytes(&key[..]).expect("malformed header name");
+            let value: Piece = value.to_vec().into();
+            headers.append(name, value);
+        }
+    };
+
+    match &data.fragments[..] {
+        [] => unreachable!("must have at least one fragment"),
+        [payload] => {
+            // TODO: propagate errors here instead of panicking
+            hpack_dec
+                .decode_with_cb(&payload[..], on_header_pair)
+                .map_err(|e| eyre::eyre!("hpack error: {e:?}"))?;
+        }
+        _ => {
+            let total_len = data.fragments.iter().map(|f| f.len()).sum();
+            // this is a slow path, let's do a little heap allocation. we could
+            // be using `RollMut` for this, but it would probably need to resize
+            // a bunch
+            let mut payload = Vec::with_capacity(total_len);
+            for frag in &data.fragments {
+                payload.extend_from_slice(&frag[..]);
+            }
+            // TODO: propagate errors here instead of panicking
+            hpack_dec
+                .decode_with_cb(&payload[..], on_header_pair)
+                .map_err(|e| eyre::eyre!("hpack error: {e:?}"))?;
+        }
+    };
+
+    match headers_or_trailers {
+        HeadersOrTrailers::Headers => {
+            // TODO: cf. https://httpwg.org/specs/rfc9113.html#HttpRequest
+            // A server SHOULD treat a request as malformed if it contains a Host header
+            // field that identifies an entity that differs from the entity in the
+            // ":authority" pseudo-header field.
+
+            // TODO: proper error handling (return 400)
+            let method = method.unwrap();
+            let scheme = scheme.unwrap();
+
+            let path = path.unwrap();
+            let path_and_query: PathAndQuery = path.parse().unwrap();
+
+            let authority = match authority {
+                Some(authority) => Some(authority),
+                None => headers
+                    .get(header::HOST)
+                    .map(|host| host.as_str().unwrap().parse().unwrap()),
+            };
+
+            let mut uri_parts: http::uri::Parts = Default::default();
+            uri_parts.scheme = Some(scheme);
+            uri_parts.authority = authority;
+            uri_parts.path_and_query = Some(path_and_query);
+
+            let uri = http::uri::Uri::from_parts(uri_parts).unwrap();
+
+            let req = Request {
+                method,
+                uri,
+                version: Version::HTTP_2,
+                headers,
+            };
+
+            let responder = Responder {
+                encoder: H2Encoder {
+                    stream_id,
+                    tx: ev_tx.clone(),
+                    state: EncoderState::ExpectResponseHeaders,
+                },
+                state: ExpectResponseHeaders,
+            };
+
+            let (piece_tx, piece_rx) = mpsc::channel::<H2BodyItem>(1); // TODO: is 1 a sensible value here?
+
+            stream_state.rx_stage = if data.end_stream {
+                StreamRxStage::Done
+            } else {
+                StreamRxStage::Body(piece_tx)
+            };
+
+            let req_body = H2Body {
+                // FIXME: that's not right. h2 requests can still specify
+                // a content-length
+                content_length: if data.end_stream { Some(0) } else { None },
+                eof: data.end_stream,
+                rx: piece_rx,
+            };
+
+            maybe_uring::spawn({
+                let driver = driver.clone();
+                async move {
+                    let mut req_body = req_body;
+                    let responder = responder;
+
+                    match driver.handle(req, &mut req_body, responder).await {
+                        Ok(_responder) => {
+                            debug!("Handler completed successfully, gave us a responder");
+                        }
+                        Err(e) => {
+                            // TODO: actually handle that error.
+                            debug!("Handler returned an error: {e}")
+                        }
+                    }
+                }
+            });
+
+            Ok(EndHeadersResult {
+                payload: EndHeadersResultPayload::Ok,
+                stream_id,
+            })
+        }
+        HeadersOrTrailers::Trailers => {
+            let mut rx_stage = StreamRxStage::Done;
+            std::mem::swap(&mut rx_stage, &mut stream_state.rx_stage);
+
+            let body_tx = match rx_stage {
+                StreamRxStage::Body(body_tx) => body_tx,
+                _ => unreachable!("trailers must follow a body"),
+            };
+
+            return Ok(EndHeadersResult {
+                stream_id,
+                payload: EndHeadersResultPayload::OkWithTrailers {
+                    body_tx,
+                    trailers: Box::new(headers),
+                },
+            });
+        }
+    }
+}
+
+#[must_use]
+struct EndHeadersResult {
+    #[allow(dead_code)]
+    stream_id: StreamId,
+    payload: EndHeadersResultPayload,
+}
+
+pub(crate) enum EndHeadersResultPayload {
+    Ok,
+    OkWithTrailers {
+        body_tx: H2BodySender,
+        trailers: Box<Headers>,
+    },
+    Err(Report),
+}
+
+impl EndHeadersResult {
+    /// Finalizes this result, asynchronously, while not
+    /// holding a lock on State.
+    pub(crate) async fn finalize(self) -> eyre::Result<()> {
+        match self.payload {
+            EndHeadersResultPayload::Ok => {
+                // well, good!
+            }
+            EndHeadersResultPayload::OkWithTrailers { body_tx, trailers } => {
+                if body_tx
+                    .send(Ok(PieceOrTrailers::Trailers(trailers)))
+                    .await
+                    .is_err()
+                {
+                    warn!("TODO: The body is being ignored, we should reset the stream");
+                }
+            }
+            EndHeadersResultPayload::Err(e) => {
+                // TODO: we should also reset the stream here, right?
+                send_goaway(&ev_tx, &state, e, KnownErrorCode::CompressionError).await;
+            }
+        }
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+pub(crate) enum HeadersOrTrailers {
+    Headers,
+    Trailers,
+}
